refactor(opentelemetry): depend on TUnit.Core instead of umbrella TUnit
Inverts the cross-assembly contract for external span ingestion so
TUnit.OpenTelemetry no longer pulls in TUnit.Engine + TUnit.Assertions
+ Microsoft.Testing.Extensions transitively.

- Move SpanData/SpanEvent/SpanLink/ReportKeyValue POCOs to TUnit.Core.
- Add ExternalSpanSink hook in TUnit.Core (Action<SpanData>? slot,
  Interlocked.CompareExchange register/unregister, first-wins).
- ActivityCollector registers IngestExternalSpan as the sink during
  Start/Stop. OtlpReceiver pushes through the sink instead of calling
  ActivityCollector.Current directly.
- Drop dead InternalsVisibleTo for TUnit.OpenTelemetry from
  TUnit.Engine.csproj.
