+++ TUnit.Engine/TUnit.Engine.csproj
@@ -9,7 +9,6 @@
   <ItemGroup>
     <InternalsVisibleTo Include="TUnit.UnitTests" />
     <InternalsVisibleTo Include="TUnit.Engine.Tests" />
-    <InternalsVisibleTo Include="TUnit.OpenTelemetry" />
     <InternalsVisibleTo Include="TUnit.OpenTelemetry.Tests" />
   </ItemGroup>
